fixing more tests
